move utils functions to different files
- common
- render
- event
