+++ src/utils/common.js
@@ -1,0 +1,20 @@
+export const getRandomInteger = (min = 0, max = 1) => {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+
+export const shuffleArray = (array) => {
+  const length = array.length - 1;
+  for (let i = length; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    [array[i], array[j]] = [array[j], array[i]];
+  }
+  return array;
+};
+
+export const getRandomBoolean = () => Math.random() >= 0.5;
+
+export const addZero = (digit) => {
+  return digit < 10
+    ? `0` + digit
+    : digit;
+};
