+++ src/view/filter.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils';
+import {createElement} from '../utils/render';
 
 const createTripFilterItemTemplate = (filter, isChecked) => {
   const {name} = filter;
