+++ src/view/sort.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils';
+import {createElement} from '../utils/render';
 
 const createSortEventsTemplate = () => {
   return `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
