+++ src/view/trip-events-list.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils';
+import {createElement} from '../utils/render';
 
 const createTripEventsListTemplate = () => {
   return `<ul class="trip-events__list"></ul>`;
